* Canon CR2: read sensor Black / White levels from MakerNotes
* Pentax K-3 Mark III Monochrome camera support
* Pentax K-3 Mark III Monochrome camera support
* Canon PowerShot SX220 HS (CHDK) camera support
* LJpeg: "ban" point transform

* upstream/develop: (69 commits)
  [CI] macOS: bump timeout for homebrew step
  Revert "Cr2Decoder: read black/white levels from ColorData4 ver.3"
  Canon PowerShot SX220 HS (CHDK) support
  CMake: fully enable `-Wshorten-64-to-32` on 32-bit builds too
  benchmarks: when ctor-ing vector with size, assume size fits `size_t`
  benchmarking: `areaToRectangle()`: take `uint64_t`, `size_t` might be 32-bit
  rstest: time is not `size_t`, now for real
  roundToMultiple()+friends: take `uint64_t`, `size_t` might be 32-bit
  PanasonicV5Decompressor: number of blocks always fits `size_t`
  PanasonicV4Decompressor: number of blocks always fits `size_t`
  roundUpDivision(): take `uint64_t`, `size_t` might be 32-bit
  PrefixCodeVectorDecoder: `extrCodeIdForLen` is never 64-bit
  ColorFilterArray: CFA size is <= 36, implicit truncation is safe
  Cr2Decoder: Canon ColorData4, ver.3: `SpecularWhiteLevel` is at `629`
  [CI] windows: *DON'T* update the MSYS2
  [CI] macOS: *DON'T* update the brew packages
  Drop obsolete Dockerfile
  copyPixelsImpl(): add padding-less special-casing back
  `copyPixels()`: ensure that we do end up with `memcpy()`
  `copyPixels()`: reimplement over a `Array2DRef`
  ...

+++ bench/librawspeed/bench/Common.h
@@ -22,9 +22,9 @@
 
 #include "rawspeedconfig.h"
 #include "adt/Point.h"
-#include <cstddef>
+#include <cstdint>
 
 bool RAWSPEED_READNONE benchmarkDryRun();
 
 rawspeed::iPoint2D RAWSPEED_READNONE
-areaToRectangle(size_t area, rawspeed::iPoint2D aspect = {2, 2});
+areaToRectangle(uint64_t area, rawspeed::iPoint2D aspect = {2, 2});
