+++ src/librawspeed/metadata/ColorFilterArray.cpp
@@ -19,6 +19,7 @@
 */
 
 #include "metadata/ColorFilterArray.h"
+#include "adt/Casts.h"
 #include "adt/Invariant.h"
 #include "adt/Point.h"
 #include "common/Common.h"
@@ -62,7 +63,7 @@ void ColorFilterArray::setSize(const iPoint2D& _size) {
   }
   if (size.area() <= 0)
     return;
-  cfa.resize(size.area());
+  cfa.resize(implicit_cast<size_t>(size.area()));
   fill(cfa.begin(), cfa.end(), CFAColor::UNKNOWN);
 }
 
@@ -99,7 +100,7 @@ void ColorFilterArray::shiftRight(int n) {
   if (n == 0)
     return;
 
-  vector<CFAColor> tmp(size.area());
+  vector<CFAColor> tmp(implicit_cast<size_t>(size.area()));
   for (int y = 0; y < size.y; ++y) {
     for (int x = 0; x < size.x; ++x) {
       tmp[x + static_cast<size_t>(y) * size.x] = getColorAt(x + n, y);
@@ -117,7 +118,7 @@ void ColorFilterArray::shiftDown(int n) {
   if (n == 0)
     return;
 
-  vector<CFAColor> tmp(size.area());
+  vector<CFAColor> tmp(implicit_cast<size_t>(size.area()));
   for (int y = 0; y < size.y; ++y) {
     for (int x = 0; x < size.x; ++x) {
       tmp[x + static_cast<size_t>(y) * size.x] = getColorAt(x, y + n);
